Adding "import" directory for conformance resources.
Also mounting "vonklog" so that the log is more accessible.  There appears
to be a bug with the resource validation URL because the FHIR resource
name is not being updated for following resources. For example, the first
resource validation is done for /Substance/$validate but sticks for the
following resources and ends up causing validation errors for a Patient that
sound be posted to /Patient/$validate but is posted to /Substance/$validate instead.