:construction: Temp add phenopackets resources to kidsfirst model
Once a bug is fixed with IG publisher we will be able to specify the
phenopackets fhir package as a dependency in the IG control file.
We can then point to a local phenopackets fhir package in the control
file or a URL where the phenopackets IG can be downloaded.
Once this is possible we can delete the phenopackets resources
from site_root/source/resources